feature(ptw): implement main functionality for ptw alias
Currently supports saving employer, check, and bonuses;
Wage and attendance are automatically tracked, increased,
and decreased according to existing logic. Almost everything
is configurable by editing constants at the top of the file.

FUNCTIONAL CHANGES:
Wages are dropped after EIGHT days, not seven; this is to provide
a small grace period that thralls would have offered, since
wage cuts previously operated on calendar days. This period can,
of course, be adjusted; eight days is just my suggestion!

TODO:
- The `adv` argument is currently NOT saved,
but probably should be! This change would require a different
way of handling the argument, though; it'd need to be passed
as a flag, instead of just being the simple `adv` arg.
- AUTOMATED PAYMENT; I need to dig through `!transaction` and
steal some logic to automate payment and report.
- Better help text
- When `!lifestyle` is updated, we can get rid of the lifestyle
arg for this command.